refactor(slice-45): move embedders and retrievers into theme packages
Group provider and search adapters under server.core.embedding and
server.core.retrieval with shim aliases so existing patches keep working
(DECISIONS #150).
