+++ docs/plan/DECISIONS.md
@@ -153,3 +153,4 @@
 | 147 | 2026-07-27 | 45 | Landed `server/core/guards/` + shim re-exports at old `server.core.*` paths; test patches retargeted to `server.core.guards.*` | Preserve caller imports for one release; mypy-safe explicit re-exports; coverage cov paths → guards | sys.modules alias shims / rewrite all callers in same PR |
 | 148 | 2026-07-27 | 45 | Land `server/core/pipeline/` (orchestrator/executors/experiment_control/startup_reconciliation); `sys.modules` shims for patch compatibility | Mega-suite patches `server.core.orchestrator.*`; function-level SLAP deferred to next PR on same branch | Split `_run_single` / bayesian in same commit as package move |
 | 149 | 2026-07-28 | 45 | Orchestrator SLAP: extract `pipeline/search.py` + `signatures.py`; keep `_update_phase` / final-status / Bayesian in `orchestrator.py` | Status helpers close over `get_storage_backend` in defining module — extracting them breaks `patch("server.core.orchestrator.get_storage_backend")`; search/signature SUTs retarget patches narrowly | Extract status helpers without patch retargeting / rewrite all mega-suite patches in one go |
+| 150 | 2026-07-28 | 45 | Land `server/core/embedding/` + `server/core/retrieval/`; `sys.modules` shims at old flat paths | Same patch-compat pattern as pipeline (#148); factory tests retarget `sys.modules` keys to `embedding.*`; rerank left flat for next PR | Move rerank in same commit / rewrite all callers without shims |
