+++ pyproject.toml
@@ -100,9 +100,9 @@ source = [
     "server/core/guards/search_index_guard.py",
     "server/core/results_analyzer.py",
     "server/models/config.py",
-    "server/core/sie_embedder.py",
+    "server/core/embedding/sie_embedder.py",
     "server/core/aim_logger.py",
-    "server/core/embedder_factory.py",
+    "server/core/embedding/embedder_factory.py",
 ]
 branch = true
