+++ server/core/embedder_factory.py
@@ -1,50 +1,13 @@
-"""Provider dispatch factory for embedding functions.
+"""Deprecated import path — use ``server.core.embedding.embedder_factory``.
 
-Returns a (embed_docs_fn, embed_query_fn) pair for the given provider string.
-The orchestrator calls get_embedder(provider) once per run and uses the returned
-functions directly — no if/elif chains in the pipeline code.
-
-Design decision (DECISIONS.md #10): factory function preferred over Protocol/ABC
-at current scale — YAGNI + Simple Design.
+Shim kept for one release (Slice 45). Uses module aliasing so existing
+``patch("server.core.embedder_factory.*")`` targets keep working.
 """
 
 from __future__ import annotations
 
-from collections.abc import Callable
+import sys
 
-EmbedDocsFn = Callable[..., list[list[float]]]
-EmbedQueryFn = Callable[[str, str], list[float]]
+from server.core.embedding import embedder_factory as _impl
 
-
-def get_embedder(provider: str) -> tuple[EmbedDocsFn, EmbedQueryFn]:
-    """Return (embed_documents_fn, embed_query_fn) for the given provider.
-
-    Args:
-        provider: One of "voyage", "local", "sie".
-
-    Returns:
-        Pair of callables with signatures:
-          embed_docs(texts: list[str], model_id: str) -> list[list[float]]
-          embed_query(text: str, model_id: str) -> list[float]
-
-    Raises:
-        ValueError: If provider is not recognised.
-    """
-    if provider == "voyage":
-        from server.core.embedder import embed_documents_voyage, embed_query_voyage
-
-        return embed_documents_voyage, embed_query_voyage
-
-    if provider == "local":
-        from server.core.local_embedder import embed_documents_local, embed_query_local
-
-        return embed_documents_local, embed_query_local
-
-    if provider == "sie":
-        from server.core.sie_embedder import embed_documents_sie, embed_query_sie
-
-        return embed_documents_sie, embed_query_sie
-
-    raise ValueError(
-        f"Unknown embedding provider '{provider}'. Supported: 'voyage', 'local', 'sie'."
-    )
+sys.modules[__name__] = _impl
