+++ server/core/embedding/__init__.py
@@ -1,0 +1,5 @@
+"""Embedding providers — Feature theme (Slice 45).
+
+Import from submodules (e.g. ``server.core.embedding.embedder_factory``).
+Temporary shims remain at ``server.core.<module>`` for one release.
+"""
