+++ server/core/retrieval/__init__.py
@@ -1,0 +1,5 @@
+"""Vector/text retrieval adapters — Feature/Behavior theme (Slice 45).
+
+Import from submodules (e.g. ``server.core.retrieval.retriever_mongo``).
+Temporary shims remain at ``server.core.<module>`` for one release.
+"""
