+++ server/utils/scope_log.py
@@ -15,12 +15,18 @@ _SCOPE_OVERRIDES: dict[str, str] = {
     "server.core.orchestrator": "Orchestrator",
     "server.core.executors": "executors",
     "server.core.startup_reconciliation": "startupReconciliation",
+    "server.core.embedding.embedder": "Embedder",
+    "server.core.embedding.local_embedder": "localEmbedder",
+    "server.core.embedding.rate_limiter": "rateLimiter",
+    "server.core.retrieval.retriever_mongo": "Retriever",
+    "server.core.retrieval.retriever_postgres": "RetrieverPostgres",
+    "server.core.reranker": "Reranker",
+    "server.core.local_reranker": "localReranker",
+    # Deprecated shim paths (Slice 45 — keep until shims removed)
     "server.core.embedder": "Embedder",
     "server.core.local_embedder": "localEmbedder",
     "server.core.retriever_mongo": "Retriever",
     "server.core.retriever_postgres": "RetrieverPostgres",
-    "server.core.reranker": "Reranker",
-    "server.core.local_reranker": "localReranker",
     "server.core.rate_limiter": "rateLimiter",
     "server.core.data_loader": "dataLoader",
     "server.core.query_loader": "queryLoader",
