+++ tests/test_postgres_sparse_hybrid.py
@@ -360,7 +360,7 @@ class TestPostgresSparseHybridFailureShould:
 
         ### When
         with (
-            patch("server.core.retriever_postgres.fetch_all", side_effect=boom),
+            patch("server.core.retrieval.retriever_postgres.fetch_all", side_effect=boom),
             caplog.at_level(logging.ERROR),
             pytest.raises(psycopg.errors.UndefinedTable),
         ):
@@ -388,7 +388,7 @@ class TestPostgresSparseHybridFailureShould:
 
         ### When
         with (
-            patch("server.core.retriever_postgres.fetch_all", side_effect=boom),
+            patch("server.core.retrieval.retriever_postgres.fetch_all", side_effect=boom),
             caplog.at_level(logging.ERROR),
             pytest.raises(psycopg.errors.UndefinedTable),
         ):
